refactor: move Topology functions to .decay module
