+++ tests/test_kinematics.py
@@ -24,7 +24,6 @@ from ampform.kinematics import (
     compute_helicity_angles,
     compute_invariant_masses,
     create_four_momentum_symbols,
-    determine_attached_final_state,
 )
 from ampform.sympy._array_expressions import ArraySlice, ArraySymbol
 
@@ -349,25 +348,6 @@ def __compute_mass(array: np.ndarray) -> np.ndarray:
     three_momentum = array[:, 1:]
     mass_squared = energy**2 - np.sum(three_momentum**2, axis=1)
     return complex_sqrt(mass_squared)
-
-
-def test_determine_attached_final_state():
-    topologies = create_isobar_topologies(4)
-    # outer states
-    for topology in topologies:
-        for i in topology.outgoing_edge_ids:
-            assert determine_attached_final_state(topology, state_id=i) == [i]
-        for i in topology.incoming_edge_ids:
-            assert determine_attached_final_state(
-                topology, state_id=i
-            ) == list(topology.outgoing_edge_ids)
-    # intermediate states
-    topology = topologies[0]
-    assert determine_attached_final_state(topology, state_id=4) == [0, 1]
-    assert determine_attached_final_state(topology, state_id=5) == [2, 3]
-    topology = topologies[1]
-    assert determine_attached_final_state(topology, state_id=4) == [1, 2, 3]
-    assert determine_attached_final_state(topology, state_id=5) == [2, 3]
 
 
 def _generate_numpy_code(expr: sp.Expr) -> str:
